backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
chromadb
sentence-transformers
python-dotenv
google-ai-generativelanguage
google-generativeai